Move agent code into an agent/ directory
This prepares the structure of the code for a client library for the
Stackdriver Debug APIs to be added in the future.
