+++ src/agent/debuglet.js
@@ -23,10 +23,10 @@ var util = require('util');
 var semver = require('semver');
 
 var v8debugapi = require('./v8debugapi.js');
-var DebugletApi = require('./debugletapi.js');
+var DebugletApi = require('../debugletapi.js');
 var scanner = require('./scanner.js');
 var Logger = require('@google/cloud-diagnostics-common').logger;
-var StatusMessage = require('./apiclasses.js').StatusMessage;
+var StatusMessage = require('../apiclasses.js').StatusMessage;
 var SourceMapper = require('./sourcemapper.js');
 
 var assert = require('assert');
