+++ src/agent/scanner.js
