+++ src/agent/v8debugapi.js
@@ -23,7 +23,7 @@
 
 /** @const */ var state = require('./state.js');
 /** @const */ var logModule = require('@google/cloud-diagnostics-common').logger;
-/** @const */ var apiclasses = require('./apiclasses.js');
+/** @const */ var apiclasses = require('../apiclasses.js');
 /** @const */ var StatusMessage = apiclasses.StatusMessage;
 
 /** @const */ var messages = {
