+++ test/standalone/test-max-data-size.js
@@ -23,9 +23,9 @@ process.env.GCLOUD_DIAGNOSTICS_CONFIG = 'test/fixtures/test-config.js';
 
 var assert = require('assert');
 var logModule = require('@google/cloud-diagnostics-common').logger;
-var v8debugapi = require('../../src/v8debugapi.js');
-var SourceMapper = require('../../src/sourcemapper.js');
-var scanner = require('../../src/scanner.js');
+var v8debugapi = require('../../src/agent/v8debugapi.js');
+var SourceMapper = require('../../src/agent/sourcemapper.js');
+var scanner = require('../../src/agent/scanner.js');
 var config = require('../../src/config.js').debug;
 var api;
